requirements.txt
streamlit>=1.28.0
scikit-learn>=1.3.0
nltk>=3.8
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
torch>=2.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
kagglehub>=0.2.0
pickle-mixin>=1.0.2
